+++ front_end/front_end_website/package.json
@@ -18,6 +18,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
+    "rsuite": "^5.40.0",
     "sweetalert2": "^11.7.31",
     "web-vitals": "^2.1.4"
   },
