+++ front_end/front_end_website/src/assets/style/GlobalStyle.css
@@ -1,1 +1,5 @@
-/*# sourceMappingURL=GlobalStyle.css.map */
+.home-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}/*# sourceMappingURL=GlobalStyle.css.map */
