+++ front_end/front_end_website/src/assets/style/GlobalStyle.scss
@@ -1,0 +1,5 @@
+.home-container {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+}
