+++ front_end/front_end_website/src/assets/style/Header.scss
@@ -1,6 +1,7 @@
 #image-shop{
     width: 150px;
     height: 130px;   
+    transform: translateX(20px);
 }
 
 #user-avatar{
