--- front_end/front_end_website/src/assets/style/Home.css
@@ -1,1 +1,0 @@
-/*# sourceMappingURL=Home.css.map */
