--- front_end/front_end_website/src/assets/style/Home.scss
@@ -1,3 +1,0 @@
-.home-container{
-    
-}
