+++ front_end/front_end_website/src/assets/style/Sidemenu.css
@@ -6,8 +6,10 @@
   padding-right: 1rem;
 }
 .sidebar .nav-link {
-  font-weight: 500;
+  font-weight: 600;
   color: white;
+  width: 100%;
+  font-size: 15px;
 }
 .sidebar .nav-link:hover {
   color: var(--bs-primary);
@@ -23,7 +25,7 @@
 #nav_accordion {
   background-color: #151a25;
   color: white;
-  height: 100%;
+  height: 562px;
   display: flex;
   justify-content: space-around;
 }/*# sourceMappingURL=Sidemenu.css.map */
