+++ front_end/front_end_website/src/assets/style/Sidemenu.scss
@@ -12,8 +12,10 @@
 	}
 
 	.nav-link {
-		font-weight: 500;
+		font-weight: 600;
 		color: white;
+		width: 100%;
+		font-size: 15px;
 	}
 
 	.nav-link:hover {
@@ -22,7 +24,6 @@
 
 	.nav-main{
 		margin-bottom: 14px;
-		// font-size: 15px;
 	}
 
 }
@@ -34,7 +35,8 @@
 #nav_accordion{
 	background-color: #151a25;
 	color: white;
-	height: 100%;
+	// height: 100%;
+	height: 562px;
 	display: flex;
 	justify-content: space-around;
 }
