+++ front_end/front_end_website/src/components/user/ModifyUserModal.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const ModifyUserModal = () => {
+  return <div></div>;
+};
+
+export default ModifyUserModal;
+
