+++ front_end/front_end_website/src/layout/DashBoard.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const DashBoard = () => {
+  return(
+    <div>hello DashBoard</div>
+  );
+};
+
+export default DashBoard;
