--- front_end/front_end_website/src/layout/Home.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-import { Container, Row, Col } from "react-bootstrap";
-
-const Home = () => {
-
-    return (
-      //   <Container className="home-container">
-      //   <Header />
-      //   <Row>
-      //     <Col xs={2}>
-      //       <SideMenu />
-      //     </Col>
-      //     <Col>
-      //       Nghĩa Pé Oniemay
-      //     </Col>
-      //   </Row>
-      // </Container>
-      <p>tét</p>
-    );
-
-};
-
-export default Home;
